Change user interface for support BigInt

+++ src/user.h
@@ -4,4 +4,4 @@
 void Intro();
 void Ending();
 int InputExp(char* dest);
-void PrintResult(long long result, int errorFlag);
+void PrintResult(BigInt result, int errorFlag);
